+++ core/src/ru/rdude/rpg/game/utils/aStar/AStarScorer.java
@@ -1,5 +1,5 @@
 package ru.rdude.rpg.game.utils.aStar;
 
 public interface AStarScorer<T extends AStarNode> {
-    double computeCost(T from, T to);
+    int computeCost(T from, T to);
 }
